*** empty log message ***

+++ PmapBgnd.c
@@ -6,7 +6,7 @@ XSetWindowBackgroundPixmap(Display * dpy, Window w, Pixmap pixmap)
 	/* FIXME... if ParentRelative is specified, then we should use our */
 	/* parent's background pixmap.  That would involve a copy???      */
 
-printf("XSetWindowBackgroundPixmap %d %d\n", w, pixmap);
+printf("XSetWindowBackgroundPixmap %d %d\n", (int)w, (int)pixmap);
 	if (pixmap == ParentRelative)
 		return 1;
